code review revisions and refactor
