+++ Gemfile
@@ -27,6 +27,3 @@ if RUBY_VERSION < "2.2.2"
   gem "rack", '<2.0'
   gem "activesupport", "<5.0"
 end
-
-# SFTP Support
-gem "net-sftp"
