+++ stockboy.gemspec
@@ -32,4 +32,5 @@ Gem::Specification.new do |s|
   s.add_runtime_dependency "httpi"
   s.add_runtime_dependency "mail"
   s.add_runtime_dependency "activesupport", ">= 3.0"
+  s.add_runtime_dependency "net-sftp"
 end
